Replace files which includes data structure definition
